requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
openai==2.7.0
packaging==25.0
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==9.0.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.44
tomli==2.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
